deleting activ admin tables

+++ app/controllers/sale_distpaches_controller.rb
@@ -10,21 +10,21 @@ class SaleDistpachesController < ApplicationController
 			@gmov = GmovDistpach.find(g[1][:id])
 			@gmov.status = ""
 			if @gmov.distpached_quantity.present?
-			  if (g[1][:distpached_quantity].to_i <= @gmov.pending_distpach.to_i )
-					@gmov.distpached_quantity += g[1][:distpached_quantity].to_i
-					@gmov.pending_distpach = @gmov.pending_distpach.to_i - g[1][:distpached_quantity].to_i
+			  if (g[1][:distpached_quantity].to_f <= @gmov.pending_distpach.to_f )
+					@gmov.distpached_quantity += g[1][:distpached_quantity].to_f
+					@gmov.pending_distpach = @gmov.pending_distpach.to_f - g[1][:distpached_quantity].to_f
 				else
 					@gmov.status = "Error"
 				end
 			else  
-				if (g[1][:distpached_quantity].to_i <= @gmov.pending_distpach.to_i )
-					@gmov.distpached_quantity = g[1][:distpached_quantity].to_i
-					@gmov.pending_distpach = @gmov.pending_distpach.to_i - g[1][:distpached_quantity].to_i
+				if (g[1][:distpached_quantity].to_i <= @gmov.pending_distpach.to_f )
+					@gmov.distpached_quantity = g[1][:distpached_quantity].to_f
+					@gmov.pending_distpach = @gmov.pending_distpach.to_f - g[1][:distpached_quantity].to_f
 				else
 					@gmov.status = "Error"
 				end
 			end
-			if(@gmov.pending_distpach.to_i == 0)
+			if(@gmov.pending_distpach.to_f == 0)
 				@gmov.status = "Completado"
 			end
 			@gmov.save!
